# public/api-spec.yaml
openapi: 3.0.3
info:
  title: CRM 시스템 API
  description: |
    **마케팅 자동화 CRM 시스템**의 백엔드 API 명세서입니다.
    
    ### 기술 스택
    - **프레임워크**: Next.js 15 (App Router)
    - **데이터베이스**: Supabase (메타데이터), MySQL (비즈니스 데이터)
    - **메시징**: CoolSMS API
    - **스케줄링**: AWS Lambda + Vercel Cron
    
    ### 중요 참고사항
    - **스케줄 필드**: `schedule_config` 사용 (legacy `schedule_settings`는 하위호환성만 지원)
    - **워크플로우**: 3단계 구조 (target_config, message_config, mapping_config)
    - **테스트 모드**: 개발 환경에서 실제 메시지 발송 없이 시뮬레이션 가능
  version: 1.0.0
  contact:
    name: CRM API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.vercel.app
    description: Production server
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "에러 메시지"
        error:
          type: string
          example: "상세 에러 정보"
    Workflow:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "리뷰 요청 워크플로우"
        description:
          type: string
          example: "고객에게 리뷰 요청 메시지를 발송하는 워크플로우"
        status:
          type: string
          enum: [draft, active, paused, archived]
          example: "active"
        trigger_type:
          type: string
          enum: [manual, schedule, webhook]
          example: "manual"
        trigger_config:
          type: object
          description: "트리거 설정"
        target_config:
          type: object
          description: "대상 설정 (manual/schedule 타입에서 사용)"
        message_config:
          type: object
          description: "메시지 설정"
        mapping_config:
          type: object
          description: "매핑 설정"
        schedule_config:
          $ref: '#/components/schemas/ScheduleConfig'
        variables:
          type: object
          description: "변수 설정"
        statistics:
          type: object
          description: "실행 통계"
        last_run_at:
          type: string
          format: date-time
          example: "2025-07-29T00:00:00Z"
          description: "마지막 실행 시간"
        next_run_at:
          type: string
          format: date-time
          example: "2025-07-30T09:00:00Z"
          description: "다음 실행 예정 시간"
        created_by:
          type: string
          example: "user"
          description: "생성자"
        created_at:
          type: string
          format: date-time
          example: "2025-07-29T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-07-29T12:00:00Z"
...
    ScheduleConfig:
      type: object
      description: "스케줄 설정 (현재 사용 중인 필드)"
      required:
        - type
        - timezone
      properties:
        type:
          type: string
          enum: [immediate, delay, scheduled, recurring]
          example: "immediate"
          description: "스케줄 실행 타입"
        timezone:
          type: string
          example: "Asia/Seoul"
          description: "시간대 설정 (모든 타입에 필수)"
        delay:
          type: integer
          description: "지연 시간 (분 단위, type이 'delay'일 때 필수)"
          example: 60
        scheduledTime:
          type: string
          description: "예약 실행 시간 (type이 'scheduled'일 때 필수)"
          example: "2025-01-30T14:35:00+09:00"
        recurringPattern:
          type: object
          description: "반복 패턴 설정 (type이 'recurring'일 때 필수)"
          required:
            - frequency
            - interval
            - time
          properties:
            frequency:
              type: string
              enum: [daily, weekly, monthly]
              example: "daily"
            interval:
              type: integer
              example: 1
              description: "반복 간격"
            time:
              type: string
              pattern: "^[0-2][0-9]:[0-5][0-9]$"
              example: "14:35"
              description: "실행 시간 (HH:MM 형식)"
            daysOfWeek:
              type: array
              items:
                type: integer
                minimum: 0
                maximum: 6
              description: "요일 설정 (0=일요일, 1=월요일, ..., weekly일 때 선택사항)"
            dayOfMonth:
              type: integer
              minimum: 1
              maximum: 31
              description: "월의 특정 일 (monthly일 때 선택사항)"
      examples:
        immediate:
          summary: "즉시 실행"
          value:
            type: "immediate"
            timezone: "Asia/Seoul"
        delay:
          summary: "지연 실행"
          value:
            type: "delay"
            delay: 60
            timezone: "Asia/Seoul"
        recurring_daily:
          summary: "매일 반복"
          value:
            type: "recurring"
            timezone: "Asia/Seoul"
            recurringPattern:
              time: "14:35"
              interval: 1
              frequency: "daily"
paths:
...
  /api/supabase/workflows/{id}:
    get:
      tags: [Workflows]
      summary: 단일 워크플로우 조회
      description: "특정 워크플로우의 상세 정보를 조회합니다."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "워크플로우 ID"
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Workflow'
        '404':
          description: 워크플로우를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Workflow not found"
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Workflows]
      summary: 워크플로우 전체 수정
      description: |
        워크플로우의 전체 또는 부분 정보를 수정합니다.
        워크플로우 타입에 따라 수정 가능한 필드가 다릅니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "워크플로우 ID"
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  description: "워크플로우 이름"
                  example: "수정된 워크플로우 이름"
                description:
                  type: string
                  description: "워크플로우 설명"
                trigger_type:
                  type: string
                  enum: [manual, schedule, webhook]
                  description: "트리거 타입"
                status:
                  type: string
                  enum: [draft, active, paused, archived]
                  description: "워크플로우 상태"
                selectedTemplates:
                  type: array
                  items:
                    type: object
                targetGroups:
                  type: array
                  items:
                    type: object
                templatePersonalizations:
                  type: object
                targetTemplateMappings:
                  type: array
                  items:
                    type: object
                schedule_config:
                  $ref: '#/components/schemas/ScheduleConfig'
                trigger_config:
                  type: object
                testSettings:
                  type: object
                steps:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Workflow'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Workflow name cannot be empty"
        '404':
          description: 워크플로우를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Workflow not found"
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Workflows]
      summary: 워크플로우 상태 변경
      description: |
        워크플로우의 상태를 변경합니다. 
        상태가 'paused' 또는 'archived'로 변경되면 관련된 스케줄 작업들이 자동으로 취소됩니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "워크플로우 ID"
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [draft, active, paused, archived]
                  description: "변경할 상태"
                  example: "active"
      responses:
        '200':
          description: 상태 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Workflow'
                  message:
                    type: string
                    examples:
                      - "워크플로우가 활성화되었습니다."
                      - "워크플로우가 일시정지되었습니다."
                      - "워크플로우가 보관되었습니다."
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    examples:
                      - "Status is required for PATCH operation"
                      - "Invalid status"
        '404':
          description: 워크플로우를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Workflow not found"
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Workflows]
      summary: 워크플로우 삭제
      description: |
        워크플로우를 영구적으로 삭제합니다.
        CASCADE 설정으로 인해 관련된 scheduled_jobs도 함께 삭제됩니다.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "워크플로우 ID"
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "워크플로우 '리뷰 요청 워크플로우'가 삭제되었습니다."
        '404':
          description: 워크플로우를 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Workflow not found"
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...